Button - refactor into smaller files (reanimated 2 part 1)
