+++ generatedTypes/components/button/ButtonConstants.d.ts
@@ -1,0 +1,21 @@
+import { ButtonSize } from './ButtonTypes';
+export declare const PADDINGS: {
+    XSMALL: number;
+    SMALL: number;
+    MEDIUM: number;
+    LARGE: number;
+};
+export declare const HORIZONTAL_PADDINGS: {
+    XSMALL: number;
+    SMALL: number;
+    MEDIUM: number;
+    LARGE: number;
+};
+export declare const MIN_WIDTH: {
+    XSMALL: number;
+    SMALL: number;
+    MEDIUM: number;
+    LARGE: number;
+};
+export declare const DEFAULT_SIZE = ButtonSize.large;
+export declare const DISABLED_COLOR: string;
