+++ generatedTypes/components/button/index.d.ts
@@ -1,146 +1,7 @@
 import React, { PureComponent } from 'react';
-import { LayoutChangeEvent, ImageStyle, TextStyle, StyleProp } from 'react-native';
-import { BaseComponentInjectedProps, ForwardRefInjectedProps, TypographyModifiers, ColorsModifiers, BackgroundColorModifier, MarginModifiers } from '../../commons/new';
-import { TouchableOpacityProps } from '../touchableOpacity';
-import { TextProps } from '../text';
-export declare enum ButtonSize {
-    xSmall = "xSmall",
-    small = "small",
-    medium = "medium",
-    large = "large"
-}
-export declare enum ButtonAnimationDirection {
-    center = "center",
-    left = "left",
-    right = "right"
-}
-export declare type ButtonProps = TouchableOpacityProps & TypographyModifiers & ColorsModifiers & BackgroundColorModifier & MarginModifiers & {
-    /**
-     * Text to show inside the button
-     */
-    label?: string;
-    /**
-     * The Button text color (inherited from Text component)
-     */
-    color?: string;
-    /**
-     * Icon image source or a callback function that returns a source
-     */
-    iconSource?: object | number | Function;
-    /**
-     * Icon image style
-     */
-    iconStyle?: StyleProp<ImageStyle>;
-    /**
-     * Should the icon be right to the label
-     */
-    iconOnRight?: boolean;
-    /**
-     * whether the icon should flip horizontally on RTL locals
-     */
-    supportRTL?: boolean;
-    /**
-     * Color of the button background
-     */
-    backgroundColor?: string;
-    /**
-     * Color of the disabled button background
-     */
-    disabledBackgroundColor?: string;
-    /**
-     * Size of the button [large, medium, small, xSmall]
-     */
-    size?: ButtonSize;
-    /**
-     * Custom border radius.
-     */
-    borderRadius?: number;
-    /**
-     * Actions handler
-     */
-    onPress?: (props: any) => void;
-    /**
-     * Disable interactions for the component
-     */
-    disabled?: boolean;
-    /**
-     * Button will have outline style
-     */
-    outline?: boolean;
-    /**
-     * The outline color
-     */
-    outlineColor?: string;
-    /**
-     * The outline width
-     */
-    outlineWidth?: number;
-    /**
-     * Button will look like a link
-     */
-    link?: boolean;
-    /**
-     * label color for when it's displayed as link
-     */
-    linkColor?: string;
-    /**
-     * Additional styles for label text
-     */
-    labelStyle?: StyleProp<TextStyle>;
-    /**
-     * Props that will be passed to the button's Text label.
-     */
-    labelProps?: TextProps;
-    /**
-     * should the button act as a coast to coast button (no border radius)
-     */
-    fullWidth?: boolean;
-    /**
-     * should the button be a round button
-     */
-    round?: boolean;
-    /**
-     * Control shadow visibility (iOS-only)
-     */
-    enableShadow?: boolean;
-    /**
-     * avoid inner button padding
-     */
-    avoidInnerPadding?: boolean;
-    /**
-     * avoid minimum width constraints
-     */
-    avoidMinWidth?: boolean;
-    /**
-     * callback for getting activeBackgroundColor (e.g. (calculatedBackgroundColor, prop) => {...})
-     * better set using ThemeManager
-     */
-    getActiveBackgroundColor?: (backgroundColor: string, props: any) => string;
-    /**
-     * should animate layout change
-     * Note?: For Android you must set 'setLayoutAnimationEnabledExperimental(true)' via RN's 'UIManager'
-     */
-    animateLayout?: boolean;
-    /**
-     * the direction of the animation ('left' and 'right' will effect the button's own alignment)
-     */
-    animateTo?: ButtonAnimationDirection;
-};
-export declare type ButtonPropTypes = ButtonProps;
-export declare type ButtonState = {
-    size?: number;
-    borderRadius?: number;
-    isLandscape?: boolean;
-};
-declare type Props = ButtonProps & BaseComponentInjectedProps & ForwardRefInjectedProps;
-/**
- * @description: Basic button component
- * @extends: TouchableOpacity
- * @extendslink: docs/TouchableOpacity
- * @modifiers: margin, background
- * @gif: https://media.giphy.com/media/xULW8j5WzsuPytqklq/giphy.gif
- * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/ButtonsScreen.tsx
- */
+import { LayoutChangeEvent, ImageStyle } from 'react-native';
+import { ButtonSize, ButtonAnimationDirection, ButtonProps, ButtonPropTypes, ButtonState, Props } from './ButtonTypes';
+export { ButtonSize, ButtonAnimationDirection, ButtonProps, ButtonPropTypes };
 declare class Button extends PureComponent<Props, ButtonState> {
     static displayName: string;
     static defaultProps: {
@@ -328,7 +189,7 @@ declare class Button extends PureComponent<Props, ButtonState> {
     } | {
         shadowColor: any;
     })[] | undefined;
-    getIconStyle(): StyleProp<ImageStyle>[];
+    getIconStyle(): import("react-native").StyleProp<ImageStyle>[];
     getAnimationDirectionStyle(): {
         alignSelf: string;
     } | undefined;
@@ -500,89 +361,32 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         flatMap: <U_3, This = undefined>(callback: (this: This, value: false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined, index: number, array: (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined)[]) => U_3 | readonly U_3[], thisArg?: This | undefined) => U_3[];
         flat: <A, D extends number = 1>(this: A, depth?: D | undefined) => FlatArray<A, D>[];
     } | null | undefined;
-    onPress?: ((props: TouchableOpacityProps) => void) | undefined;
+    onPress?: ((props: import("../touchableOpacity").TouchableOpacityProps) => void) | undefined;
 } & import("../../commons/modifiers").CustomModifier & {
-    /**
-     * Text to show inside the button
-     */
     label?: string | undefined;
-    /**
-     * The Button text color (inherited from Text component)
-     */
     color?: string | undefined;
-    /**
-     * Icon image source or a callback function that returns a source
-     */
     iconSource?: number | object | Function | undefined;
-    /**
-     * Icon image style
-     */
-    iconStyle?: StyleProp<ImageStyle>;
-    /**
-     * Should the icon be right to the label
-     */
+    iconStyle?: import("react-native").StyleProp<ImageStyle>;
     iconOnRight?: boolean | undefined;
-    /**
-     * whether the icon should flip horizontally on RTL locals
-     */
     supportRTL?: boolean | undefined;
-    /**
-     * Color of the button background
-     */
     backgroundColor?: string | undefined;
-    /**
-     * Color of the disabled button background
-     */
     disabledBackgroundColor?: string | undefined;
-    /**
-     * Size of the button [large, medium, small, xSmall]
-     */
     size?: ButtonSize | undefined;
-    /**
-     * Custom border radius.
-     */
     borderRadius?: number | undefined;
-    /**
-     * Actions handler
-     */
     onPress?: ((props: any) => void) | undefined;
-    /**
-     * Disable interactions for the component
-     */
     disabled?: boolean | undefined;
-    /**
-     * Button will have outline style
-     */
     outline?: boolean | undefined;
-    /**
-     * The outline color
-     */
     outlineColor?: string | undefined;
-    /**
-     * The outline width
-     */
     outlineWidth?: number | undefined;
-    /**
-     * Button will look like a link
-     */
     link?: boolean | undefined;
-    /**
-     * label color for when it's displayed as link
-     */
     linkColor?: string | undefined;
-    /**
-     * Additional styles for label text
-     */
-    labelStyle?: StyleProp<TextStyle>;
-    /**
-     * Props that will be passed to the button's Text label.
-     */
+    labelStyle?: import("react-native").StyleProp<import("react-native").TextStyle>;
     labelProps?: (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
         color?: string | undefined;
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -590,7 +394,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -598,7 +402,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -606,43 +410,17 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | undefined;
-    /**
-     * should the button act as a coast to coast button (no border radius)
-     */
     fullWidth?: boolean | undefined;
-    /**
-     * should the button be a round button
-     */
     round?: boolean | undefined;
-    /**
-     * Control shadow visibility (iOS-only)
-     */
     enableShadow?: boolean | undefined;
-    /**
-     * avoid inner button padding
-     */
     avoidInnerPadding?: boolean | undefined;
-    /**
-     * avoid minimum width constraints
-     */
     avoidMinWidth?: boolean | undefined;
-    /**
-     * callback for getting activeBackgroundColor (e.g. (calculatedBackgroundColor, prop) => {...})
-     * better set using ThemeManager
-     */
     getActiveBackgroundColor?: ((backgroundColor: string, props: any) => string) | undefined;
-    /**
-     * should animate layout change
-     * Note?: For Android you must set 'setLayoutAnimationEnabledExperimental(true)' via RN's 'UIManager'
-     */
     animateLayout?: boolean | undefined;
-    /**
-     * the direction of the animation ('left' and 'right' will effect the button's own alignment)
-     */
     animateTo?: ButtonAnimationDirection | undefined;
 } & {
     useCustomTheme?: boolean | undefined;
@@ -809,89 +587,32 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         flatMap: <U_3, This = undefined>(callback: (this: This, value: false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined, index: number, array: (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined)[]) => U_3 | readonly U_3[], thisArg?: This | undefined) => U_3[];
         flat: <A, D extends number = 1>(this: A, depth?: D | undefined) => FlatArray<A, D>[];
     } | null | undefined;
-    onPress?: ((props: TouchableOpacityProps) => void) | undefined;
+    onPress?: ((props: import("../touchableOpacity").TouchableOpacityProps) => void) | undefined;
 } & import("../../commons/modifiers").CustomModifier & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & {
-    /**
-     * Text to show inside the button
-     */
     label?: string | undefined;
-    /**
-     * The Button text color (inherited from Text component)
-     */
     color?: string | undefined;
-    /**
-     * Icon image source or a callback function that returns a source
-     */
     iconSource?: number | object | Function | undefined;
-    /**
-     * Icon image style
-     */
-    iconStyle?: StyleProp<ImageStyle>;
-    /**
-     * Should the icon be right to the label
-     */
+    iconStyle?: import("react-native").StyleProp<ImageStyle>;
     iconOnRight?: boolean | undefined;
-    /**
-     * whether the icon should flip horizontally on RTL locals
-     */
     supportRTL?: boolean | undefined;
-    /**
-     * Color of the button background
-     */
     backgroundColor?: string | undefined;
-    /**
-     * Color of the disabled button background
-     */
     disabledBackgroundColor?: string | undefined;
-    /**
-     * Size of the button [large, medium, small, xSmall]
-     */
     size?: ButtonSize | undefined;
-    /**
-     * Custom border radius.
-     */
     borderRadius?: number | undefined;
-    /**
-     * Actions handler
-     */
     onPress?: ((props: any) => void) | undefined;
-    /**
-     * Disable interactions for the component
-     */
     disabled?: boolean | undefined;
-    /**
-     * Button will have outline style
-     */
     outline?: boolean | undefined;
-    /**
-     * The outline color
-     */
     outlineColor?: string | undefined;
-    /**
-     * The outline width
-     */
     outlineWidth?: number | undefined;
-    /**
-     * Button will look like a link
-     */
     link?: boolean | undefined;
-    /**
-     * label color for when it's displayed as link
-     */
     linkColor?: string | undefined;
-    /**
-     * Additional styles for label text
-     */
-    labelStyle?: StyleProp<TextStyle>;
-    /**
-     * Props that will be passed to the button's Text label.
-     */
+    labelStyle?: import("react-native").StyleProp<import("react-native").TextStyle>;
     labelProps?: (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
         color?: string | undefined;
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -899,7 +620,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -907,7 +628,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -915,43 +636,17 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | undefined;
-    /**
-     * should the button act as a coast to coast button (no border radius)
-     */
     fullWidth?: boolean | undefined;
-    /**
-     * should the button be a round button
-     */
     round?: boolean | undefined;
-    /**
-     * Control shadow visibility (iOS-only)
-     */
     enableShadow?: boolean | undefined;
-    /**
-     * avoid inner button padding
-     */
     avoidInnerPadding?: boolean | undefined;
-    /**
-     * avoid minimum width constraints
-     */
     avoidMinWidth?: boolean | undefined;
-    /**
-     * callback for getting activeBackgroundColor (e.g. (calculatedBackgroundColor, prop) => {...})
-     * better set using ThemeManager
-     */
     getActiveBackgroundColor?: ((backgroundColor: string, props: any) => string) | undefined;
-    /**
-     * should animate layout change
-     * Note?: For Android you must set 'setLayoutAnimationEnabledExperimental(true)' via RN's 'UIManager'
-     */
     animateLayout?: boolean | undefined;
-    /**
-     * the direction of the animation ('left' and 'right' will effect the button's own alignment)
-     */
     animateTo?: ButtonAnimationDirection | undefined;
 } & {
     useCustomTheme?: boolean | undefined;
@@ -1118,89 +813,32 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         flatMap: <U_3, This = undefined>(callback: (this: This, value: false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined, index: number, array: (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined)[]) => U_3 | readonly U_3[], thisArg?: This | undefined) => U_3[];
         flat: <A, D extends number = 1>(this: A, depth?: D | undefined) => FlatArray<A, D>[];
     } | null | undefined;
-    onPress?: ((props: TouchableOpacityProps) => void) | undefined;
+    onPress?: ((props: import("../touchableOpacity").TouchableOpacityProps) => void) | undefined;
 } & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & import("../../commons/modifiers").CustomModifier & {
-    /**
-     * Text to show inside the button
-     */
     label?: string | undefined;
-    /**
-     * The Button text color (inherited from Text component)
-     */
     color?: string | undefined;
-    /**
-     * Icon image source or a callback function that returns a source
-     */
     iconSource?: number | object | Function | undefined;
-    /**
-     * Icon image style
-     */
-    iconStyle?: StyleProp<ImageStyle>;
-    /**
-     * Should the icon be right to the label
-     */
+    iconStyle?: import("react-native").StyleProp<ImageStyle>;
     iconOnRight?: boolean | undefined;
-    /**
-     * whether the icon should flip horizontally on RTL locals
-     */
     supportRTL?: boolean | undefined;
-    /**
-     * Color of the button background
-     */
     backgroundColor?: string | undefined;
-    /**
-     * Color of the disabled button background
-     */
     disabledBackgroundColor?: string | undefined;
-    /**
-     * Size of the button [large, medium, small, xSmall]
-     */
     size?: ButtonSize | undefined;
-    /**
-     * Custom border radius.
-     */
     borderRadius?: number | undefined;
-    /**
-     * Actions handler
-     */
     onPress?: ((props: any) => void) | undefined;
-    /**
-     * Disable interactions for the component
-     */
     disabled?: boolean | undefined;
-    /**
-     * Button will have outline style
-     */
     outline?: boolean | undefined;
-    /**
-     * The outline color
-     */
     outlineColor?: string | undefined;
-    /**
-     * The outline width
-     */
     outlineWidth?: number | undefined;
-    /**
-     * Button will look like a link
-     */
     link?: boolean | undefined;
-    /**
-     * label color for when it's displayed as link
-     */
     linkColor?: string | undefined;
-    /**
-     * Additional styles for label text
-     */
-    labelStyle?: StyleProp<TextStyle>;
-    /**
-     * Props that will be passed to the button's Text label.
-     */
+    labelStyle?: import("react-native").StyleProp<import("react-native").TextStyle>;
     labelProps?: (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
         color?: string | undefined;
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1208,7 +846,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1216,7 +854,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1224,43 +862,17 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | undefined;
-    /**
-     * should the button act as a coast to coast button (no border radius)
-     */
     fullWidth?: boolean | undefined;
-    /**
-     * should the button be a round button
-     */
     round?: boolean | undefined;
-    /**
-     * Control shadow visibility (iOS-only)
-     */
     enableShadow?: boolean | undefined;
-    /**
-     * avoid inner button padding
-     */
     avoidInnerPadding?: boolean | undefined;
-    /**
-     * avoid minimum width constraints
-     */
     avoidMinWidth?: boolean | undefined;
-    /**
-     * callback for getting activeBackgroundColor (e.g. (calculatedBackgroundColor, prop) => {...})
-     * better set using ThemeManager
-     */
     getActiveBackgroundColor?: ((backgroundColor: string, props: any) => string) | undefined;
-    /**
-     * should animate layout change
-     * Note?: For Android you must set 'setLayoutAnimationEnabledExperimental(true)' via RN's 'UIManager'
-     */
     animateLayout?: boolean | undefined;
-    /**
-     * the direction of the animation ('left' and 'right' will effect the button's own alignment)
-     */
     animateTo?: ButtonAnimationDirection | undefined;
 } & {
     useCustomTheme?: boolean | undefined;
@@ -1427,89 +1039,32 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         flatMap: <U_3, This = undefined>(callback: (this: This, value: false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined, index: number, array: (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | import("react-native").RecursiveArray<false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined> | readonly (false | import("react-native").ViewStyle | import("react-native").RegisteredStyle<import("react-native").ViewStyle> | null | undefined)[] | null | undefined)[]) => U_3 | readonly U_3[], thisArg?: This | undefined) => U_3[];
         flat: <A, D extends number = 1>(this: A, depth?: D | undefined) => FlatArray<A, D>[];
     } | null | undefined;
-    onPress?: ((props: TouchableOpacityProps) => void) | undefined;
+    onPress?: ((props: import("../touchableOpacity").TouchableOpacityProps) => void) | undefined;
 } & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & {
-    /**
-     * Text to show inside the button
-     */
     label?: string | undefined;
-    /**
-     * The Button text color (inherited from Text component)
-     */
     color?: string | undefined;
-    /**
-     * Icon image source or a callback function that returns a source
-     */
     iconSource?: number | object | Function | undefined;
-    /**
-     * Icon image style
-     */
-    iconStyle?: StyleProp<ImageStyle>;
-    /**
-     * Should the icon be right to the label
-     */
+    iconStyle?: import("react-native").StyleProp<ImageStyle>;
     iconOnRight?: boolean | undefined;
-    /**
-     * whether the icon should flip horizontally on RTL locals
-     */
     supportRTL?: boolean | undefined;
-    /**
-     * Color of the button background
-     */
     backgroundColor?: string | undefined;
-    /**
-     * Color of the disabled button background
-     */
     disabledBackgroundColor?: string | undefined;
-    /**
-     * Size of the button [large, medium, small, xSmall]
-     */
     size?: ButtonSize | undefined;
-    /**
-     * Custom border radius.
-     */
     borderRadius?: number | undefined;
-    /**
-     * Actions handler
-     */
     onPress?: ((props: any) => void) | undefined;
-    /**
-     * Disable interactions for the component
-     */
     disabled?: boolean | undefined;
-    /**
-     * Button will have outline style
-     */
     outline?: boolean | undefined;
-    /**
-     * The outline color
-     */
     outlineColor?: string | undefined;
-    /**
-     * The outline width
-     */
     outlineWidth?: number | undefined;
-    /**
-     * Button will look like a link
-     */
     link?: boolean | undefined;
-    /**
-     * label color for when it's displayed as link
-     */
     linkColor?: string | undefined;
-    /**
-     * Additional styles for label text
-     */
-    labelStyle?: StyleProp<TextStyle>;
-    /**
-     * Props that will be passed to the button's Text label.
-     */
+    labelStyle?: import("react-native").StyleProp<import("react-native").TextStyle>;
     labelProps?: (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
         color?: string | undefined;
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & import("../../commons/modifiers").CustomModifier & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1517,7 +1072,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & import("../../commons/modifiers").CustomModifier & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1525,7 +1080,7 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | (import("react-native").TextProps & Partial<Record<"text10" | "text20" | "text30" | "text40" | "text50" | "text60" | "text65" | "text70" | "text80" | "text90" | "text100" | "text10T" | "text10L" | "text10R" | "text10M" | "text10BO" | "text10H" | "text10BL" | "text20T" | "text20L" | "text20R" | "text20M" | "text20BO" | "text20H" | "text20BL" | "text30T" | "text30L" | "text30R" | "text30M" | "text30BO" | "text30H" | "text30BL" | "text40T" | "text40L" | "text40R" | "text40M" | "text40BO" | "text40H" | "text40BL" | "text50T" | "text50L" | "text50R" | "text50M" | "text50BO" | "text50H" | "text50BL" | "text60T" | "text60L" | "text60R" | "text60M" | "text60BO" | "text60H" | "text60BL" | "text65T" | "text65L" | "text65R" | "text65M" | "text65BO" | "text65H" | "text65BL" | "text70T" | "text70L" | "text70R" | "text70M" | "text70BO" | "text70H" | "text70BL" | "text80T" | "text80L" | "text80R" | "text80M" | "text80BO" | "text80H" | "text80BL" | "text90T" | "text90L" | "text90R" | "text90M" | "text90BO" | "text90H" | "text90BL" | "text100T" | "text100L" | "text100R" | "text100M" | "text100BO" | "text100H" | "text100BL", boolean>> & Partial<Record<"black" | "white" | "dark10" | "dark20" | "dark30" | "dark40" | "dark50" | "dark60" | "dark70" | "dark80" | "grey10" | "grey20" | "grey30" | "grey40" | "grey50" | "grey60" | "grey70" | "grey80" | "blue10" | "blue20" | "blue30" | "blue40" | "blue50" | "blue60" | "blue70" | "blue80" | "cyan10" | "cyan20" | "cyan30" | "cyan40" | "cyan50" | "cyan60" | "cyan70" | "cyan80" | "green10" | "green20" | "green30" | "green40" | "green50" | "green60" | "green70" | "green80" | "yellow10" | "yellow20" | "yellow30" | "yellow40" | "yellow50" | "yellow60" | "yellow70" | "yellow80" | "orange10" | "orange20" | "orange30" | "orange40" | "orange50" | "orange60" | "orange70" | "orange80" | "red10" | "red20" | "red30" | "red40" | "red50" | "red60" | "red70" | "red80" | "purple10" | "purple20" | "purple30" | "purple40" | "purple50" | "purple60" | "purple70" | "purple80" | "violet10" | "violet20" | "violet30" | "violet40" | "violet50" | "violet60" | "violet70" | "violet80", boolean>> & Partial<Record<"margin" | "marginL" | "marginT" | "marginR" | "marginB" | "marginH" | "marginV", boolean>> & {
@@ -1533,43 +1088,17 @@ declare const _default: React.ComponentClass<(Pick<import("react-native").Toucha
         center?: boolean | undefined;
         uppercase?: boolean | undefined;
         highlightString?: string | undefined;
-        highlightStyle?: TextStyle | undefined;
+        highlightStyle?: import("react-native").TextStyle | undefined;
         animated?: boolean | undefined;
         textAlign?: string | undefined;
     }) | undefined;
-    /**
-     * should the button act as a coast to coast button (no border radius)
-     */
     fullWidth?: boolean | undefined;
-    /**
-     * should the button be a round button
-     */
     round?: boolean | undefined;
-    /**
-     * Control shadow visibility (iOS-only)
-     */
     enableShadow?: boolean | undefined;
-    /**
-     * avoid inner button padding
-     */
     avoidInnerPadding?: boolean | undefined;
-    /**
-     * avoid minimum width constraints
-     */
     avoidMinWidth?: boolean | undefined;
-    /**
-     * callback for getting activeBackgroundColor (e.g. (calculatedBackgroundColor, prop) => {...})
-     * better set using ThemeManager
-     */
     getActiveBackgroundColor?: ((backgroundColor: string, props: any) => string) | undefined;
-    /**
-     * should animate layout change
-     * Note?: For Android you must set 'setLayoutAnimationEnabledExperimental(true)' via RN's 'UIManager'
-     */
     animateLayout?: boolean | undefined;
-    /**
-     * the direction of the animation ('left' and 'right' will effect the button's own alignment)
-     */
     animateTo?: ButtonAnimationDirection | undefined;
 } & {
     useCustomTheme?: boolean | undefined;
